# Improve crypto performance