@yaakov-h 2020-09-18
Merge conflicts, tests, and ideally I'd rather use `Span<>` than `ArraySegment<>` but that means either taking a package dependency or dropping <= netstandard2.1 I think, and I'm not sure which way we want to go on that overall.